Added maxDelayForGCTimers option in seconds to limit GC timers.
Used for full and eden timers.
If 0 means no limit as it was before.
Useful when timers were set to inf and GC is not triggered anymore.
Limiting the timers makes sure that GC will be forced and new delay will be calculated.
Otherwise will need to wait for memory pressure or exceeding max heap size to trigger GC.